Update for Deliverable 5
Major change is that the whole of the fighting widget is complete, though still needs bug tested in case. All buttons are working for fighting, with an added log on top of the screen to see who has done what recently. Enemy still lacks healing capabilities, but can otherwise fight. Fighting still needs an "end", but can get to it quickly.