Progress for Deliverable 6
Has some updates, though I've been falling behind somewhat. Fixed up the fighting widget some, added a pause menu, and made the chest actually give you the loot it generates. Also included the sprite sheet as well as the guide for what each sprite is intended to be. it's not very clean looking, but it's the best I could manage for what I had intended.